4_Intermediate_Interface_V6.1
Minor Update, saving now includes the window size,
Backlog and project summary updated

+++ Projects/PracticeProjects/WPF_Beta_Interface/Create_And_Move_Zones/Create_And_Move_Zones/Views/CustomEvents/DropApplicationAndResizeChain/DeskTopDroppedEventArgs.cs
@@ -17,7 +17,7 @@ namespace Create_And_Move_Zones.Views.CustomEvents.DropApplicationAndResizeChain
         public double X { get; set; }
         public double Y { get; set; }
 
-        public int ZoneNumber { get; set; }
+        public short ZoneNumber { get; set; }
 
 
     } //end of class
